Add notification model superclass.
